+++ nagra/writer.py
@@ -157,7 +157,11 @@ class WriterMixin:
         rows = df[self.columns].values
         return self.executemany(rows)
 
-    def from_polars(self, df: "LazyFrame"):
+    def from_polars(self, df: "LazyFrame", batch: bool = False):
+        """
+        Write data from a polars LazyFrame. Set `batch` to True
+        to enable streaming through the collect_batches() method
+        """
         from polars import Struct, col
 
         # Ignore extra columns
@@ -166,13 +170,15 @@ class WriterMixin:
         # Convert non-basic types to string
         for name, dtype in schema.items():
             if dtype == Struct:
-                df = df.with_columns(col("json").struct.json_encode())
+                df = df.with_columns(col(name).struct.json_encode())
+
+        if batch:
+            chunks = df.collect_batches()
+        else:
+            chunks = [df.collect()]
 
         res = []
-        for start, stop in _slicer():
-            chunk = df.slice(start, stop).collect()
-            if chunk.is_empty():
-                break
+        for chunk in chunks:
             rows = chunk.iter_rows()
             res += self.executemany(rows)
         return res
@@ -184,7 +190,10 @@ class WriterMixin:
 
         # Extract dict values - allows for field names or dotted column format
         f_or_c = list(zip(field_names, self.columns))
-        rows = (tuple(getter(record, field, col) for col, field in f_or_c) for record in records)
+        rows = (
+            tuple(getter(record, field, col) for col, field in f_or_c)
+            for record in records
+        )
         return self.executemany(rows)
 
 
@@ -198,6 +207,7 @@ def getter(record, field, col):
     if col in record:
         return record[col]
     raise KeyError(f"KeyError: neither {field} or {col} found")
+
 
 def _slicer(chunk_size=10_000):
     start = 0